+++ packages/mobx-core/index.ts
@@ -1,0 +1,24 @@
+/**
+ * (c) Michel Weststrate 2015 - 2016
+ * MIT Licensed
+ *
+ * Welcome to the mobx sources! To get an global overview of how MobX internally works,
+ * this is a good place to start:
+ * https://medium.com/@mweststrate/becoming-fully-reactive-an-in-depth-explanation-of-mobservable-55995262a254#.xvbh6qd74
+ *
+ */
+
+ // TODO: document all, and generate docs!
+ export {createAction, runInAction, useStrict} from "./core/action"
+ export {Atom, isAtom} from "./core/atom"
+
+ export {computed, isComputedValue} from "./core/computedvalue"
+ export {untracked, isComputingDerivation} from "./core/derivation"
+ export {MobxState} from "./core/mobxstate"
+ export {cell, isObservableValue} from "./core/observablevalue"
+ export {Reaction, onReactionError} from "./core/reaction"
+ export {autorun, reaction, when} from "./utils/autorun"
+ export {defaultComparer, identityComparer} from "./utils/comparer"
+ export {getDependencyTree, getObserverTree} from "./utils/extras"
+ export {transaction} from "./utils/transaction"
+ export * from "./utils/utils";