+++ packages/mobx-core/utils/autorun.ts
@@ -1,91 +1,76 @@
 import { Lambda, invariant, fail, EMPTY_OBJECT } from "./utils"
-import { isModifierDescriptor } from "../types/modifiers"
 import { Reaction, IReactionPublic, IReactionDisposer } from "../core/reaction"
 import { untrackedStart, untrackedEnd } from "../core/derivation"
-import { isAction, runInAction, createAction } from "./action"
-import { IEqualsComparer, comparer } from "./comparer"
+import { isAction, runInAction, createAction } from "../core/action"
+import { IEqualsComparer, defaultComparer } from "./comparer"
 import { getMessage } from "../utils/messages"
-import { MobxState } from "./mobxstate";
+import { MobxState } from "../core/mobxstate";
+
+export interface IAutorunOptions {
+    delay?: number
+    name?: string
+}
 
 /**
  * Creates a named reactive view and keeps it alive, so that the view is always
  * updated if one of the dependencies changes, even when the view is not further used by something else.
- * @param name The view name
  * @param view The reactive view
  * @returns disposer function, which can be used to stop the view from being updated in the future.
  */
 export function autorun(
     context: MobxState,
-    name: string =  "Autorun@" + (context.nextId()),
     view: (r: IReactionPublic) => any,
+    opts?: IAutorunOptions
 ): IReactionDisposer {
     invariant(typeof view === "function", getMessage("m004"))
     invariant(isAction(view) === false, getMessage("m005"))
 
-    const reaction = new Reaction(context, name, function() {
-        this.track(reactionRunner)
-    })
+    const name: string = (opts && opts.name) || "Autorun@" + (context.nextId())
+    const delay = (opts && opts.delay) || 0
+    let reaction: Reaction;
+
+    if (delay === 0) {
+        // normal autorun
+        reaction = new Reaction(context, name, function(this: Reaction) {
+            this.track(reactionRunner)
+        })
+    } else {
+        // debounced autorun
+        let isScheduled = false
+
+        reaction = new Reaction(context, name, () => {
+            if (!isScheduled) {
+                isScheduled = true
+                setTimeout(() => {
+                    isScheduled = false
+                    if (!reaction.isDisposed) reaction.track(reactionRunner)
+                }, delay)
+            }
+        })
+    }
 
     function reactionRunner() {
         view(reaction)
     }
 
     reaction.schedule()
-
     return reaction.getDisposer()
 }
 
-/**
- * Similar to 'observer', observes the given predicate until it returns true.
- * Once it returns true, the 'effect' function is invoked an the observation is cancelled.
- * @param name
- * @param predicate
- * @param effect
- * @returns disposer function to prematurely end the observer.
- */
 export function when(
     context: MobxState,
-    name: string,
     predicate: () => boolean,
     effect: Lambda,
+    opts?: { name?: string }
 ): IReactionDisposer {
-    return autorun(context, name, r => {
+    return autorun(context, r => {
         if (predicate()) {
             r.dispose()
             const prevUntracked = untrackedStart(context)
             effect();
             untrackedEnd(context, prevUntracked)
         }
-    })
-}
-
-export function autorunAsync(
-    context: MobxState,
-    name: string,
-    func: (r: IReactionPublic) => any,
-    delay?: number,
-): IReactionDisposer {
-    invariant(isAction(func) === false, getMessage("m006"))
-    if (delay === void 0) delay = 1
-
-    let isScheduled = false
-
-    const r = new Reaction(context, name, () => {
-        if (!isScheduled) {
-            isScheduled = true
-            setTimeout(() => {
-                isScheduled = false
-                if (!r.isDisposed) r.track(reactionRunner)
-            }, delay)
-        }
-    })
-
-    function reactionRunner() {
-        func(r)
-    }
-
-    r.schedule()
-    return r.getDisposer()
+    }, opts)
 }
 
 export interface IReactionOptions {
@@ -98,12 +83,6 @@ export interface IReactionOptions {
     name?: string
 }
 
-/**
- *
- * Basically sugar for computed(expr).observe(action(effect))
- * or
- * autorun(() => action(effect)(expr));
- */
 export function reaction<T>(
     context: MobxState,
     expression: (r: IReactionPublic) => T,
@@ -115,9 +94,6 @@ export function reaction<T>(
     }
     if (arguments.length > 4) {
         fail(getMessage("m007"))
-    }
-    if (isModifierDescriptor(expression)) {
-        fail(getMessage("m008"))
     }
 
     const name = opts.name || "Reaction@" + (context.nextId())
@@ -131,7 +107,7 @@ export function reaction<T>(
     let isScheduled = false
     let value: T
 
-    const equals = opts.equals || comparer.default
+    const equals = opts.equals || defaultComparer
 
     const r = new Reaction(context, name, () => {
         if (firstTime || (delay) < 1) {